refactor(string_additional_calculator): 커스텀 구분자 구현
